wip: handling users, dashboards, queries somewhat
noticing that attempting to mirror clients to perform the saving of data
for us is not as straight forward as it might be as sub-apis, return
different data sets to internal apis. e.g. api/users does not include
api_key so it needs to be fetched with api/users/<id>.
So committing what we have before changing direction for file client.